Added support for Module Folders output.
NPM introduced support for in-package typings,
it is done by adding typings key into package.json file, which should refer to the
typings file.
when this is the case, the d.ts file is threated as module folder, and declare module
is not allowed.

options `outputAsModuleFolder` was added to add support for module
folder output.